+++ ArmsReach/arms-reach/scripts/main.js
@@ -1,3 +1,20 @@
+class ArmsReachVariables
+{
+  static door_interaction_lastTime = 0;
+  static door_interaction_keydown = false;
+  static door_interaction_cameraCentered = false;
+  
+  static lastData = {
+    x: 0.0,
+    y: 0.0,
+    up: 0,
+    down: 0,
+    left: 0,
+    right: 0
+  };
+}
+
+
 // Sets the global maximum interaction distance
 Hooks.once('init', () => {
   // Global interaction distance control. Replaces prototype function of DoorControl. Danger...
@@ -9,14 +26,19 @@ Hooks.once('init', () => {
         let character = getFirstPlayerToken();
         
         if( !character ) {
-          iteractionFailNotification("No character is selected to interact with a door.");
+          iteractionFailNotification("No character is selected to interact with a door");
           return;
         }
         
-        let dist = getManhattanBetween(this, character);
+        let dist = getManhattanBetween(this, getTokenCenter(character));
         let gridSize = canvas.dimensions.size;
 
-        if ( (dist / gridSize) > game.settings.get("arms-reach", "globalInteractionDistance") ) return;
+        if ( (dist / gridSize) > game.settings.get("arms-reach", "globalInteractionDistance") ) {
+          var tokenName = getCharacterName(character);
+          if (tokenName) iteractionFailNotification("Door not within " + tokenName + "'s reach" );
+          else iteractionFailNotification("Door not in reach" );
+          return;
+        }
       }
 
       // Call original method
@@ -26,24 +48,16 @@ Hooks.once('init', () => {
 });
 
 // Door interaction
-let door_interaction_lastTime = 0;
-let door_interaction_keydown = false;
-let door_interaction_cameraCentered = false;
-
 document.addEventListener('keydown', evt => {
 	if (evt.key === 'e') {
     if(!game.settings.get("arms-reach", "hotkeyDoorInteractionCenter")) { return; }
-    if(door_interaction_cameraCentered) { return; }
+    if(ArmsReachVariables.door_interaction_cameraCentered) { return; }
     
-    if(   !document.activeElement || 
-      !document.activeElement.attributes ||
-      !document.activeElement.attributes['class'] ||
-      document.activeElement.attributes['class'].value !== "vtt game" 
-    ) { return; }
+    if( !isFocusOnCanvas() ) { return; }
     
-    if( door_interaction_keydown == false ) {
-      door_interaction_lastTime = Date.now();
-      door_interaction_keydown = true;
+    if( ArmsReachVariables.door_interaction_keydown == false ) {
+      ArmsReachVariables.door_interaction_lastTime = Date.now();
+      ArmsReachVariables.door_interaction_keydown = true;
     } else {
       // Center camera on character (if  key was pressed for a time)
       let diff = Date.now() - door_interaction_lastTime;
@@ -55,7 +69,7 @@ document.addEventListener('keydown', evt => {
           return;
         }
         
-        door_interaction_cameraCentered = true;
+        ArmsReachVariables.door_interaction_cameraCentered = true;
         canvas.animatePan({x: character.x, y: character.y});
       }
     }
@@ -64,18 +78,14 @@ document.addEventListener('keydown', evt => {
 
 document.addEventListener('keyup', evt => {
 	if (evt.key === 'e') {
-    door_interaction_keydown = false;
+    ArmsReachVariables.door_interaction_keydown = false;
 
-    if(   !document.activeElement || 
-          !document.activeElement.attributes ||
-          !document.activeElement.attributes['class'] ||
-          document.activeElement.attributes['class'].value !== "vtt game" 
-      ) { return; }
-    
-    if(door_interaction_cameraCentered) {
-      door_interaction_cameraCentered = false;
+    if(ArmsReachVariables.door_interaction_cameraCentered) {
+      ArmsReachVariables.door_interaction_cameraCentered = false;
       return;
     }
+    
+    if( !isFocusOnCanvas() ) { return; }
     
     if (!game.settings.get("arms-reach", "hotkeyDoorInteraction")) return;   
     
@@ -86,8 +96,54 @@ document.addEventListener('keyup', evt => {
       iteractionFailNotification("No character is selected to interact with a door.");
       return;
     }
+    
+    interactWithNearestDoor(character,0,0);
+	}
+});
 
-    // Distance calculation reqs.
+// Double Tap to open nearest door -------------------------------------------------
+document.addEventListener('keyup', evt => {
+	if (evt.key === 'ArrowUp' || evt.key === 'w') {
+    if(!game.settings.get("arms-reach", "hotkeyDoorInteractionDT")) { return; }
+    ifStuckInteract('up', 0, -0.5);
+  }
+  
+	if (evt.key === 'ArrowDown' || evt.key === 's') {
+    if(!game.settings.get("arms-reach", "hotkeyDoorInteractionDT")) { return; }
+    ifStuckInteract('down', 0, +0.5);
+  }
+  
+	if (evt.key === 'ArrowRight' || evt.key === 'd') {
+    if(!game.settings.get("arms-reach", "hotkeyDoorInteractionDT")) { return; }
+    ifStuckInteract('right', +0.5, 0);
+  }
+  
+	if (evt.key === 'ArrowLeft' || evt.key === 'a') {
+    if(!game.settings.get("arms-reach", "hotkeyDoorInteractionDT")) { return; }
+    ifStuckInteract('left', -0.5, 0);
+  }
+});
+
+function ifStuckInteract(key, offsetx, offsety) {
+  let character = getFirstPlayerToken();
+  if(!character) return;
+  
+  if( Date.now() - ArmsReachVariables.lastData[key] > 300 ) {
+    ArmsReachVariables.lastData.x = character.x;
+    ArmsReachVariables.lastData.y = character.y;
+    ArmsReachVariables.lastData[key] = Date.now();
+    return;
+  }
+  
+  // See if character is stuck
+  if(character.x == ArmsReachVariables.lastData.x && character.y == ArmsReachVariables.lastData.y) {
+    interactWithNearestDoor(character, offsetx, offsety);
+  }
+}
+
+// Interact with door ------------------------------------------------------------------
+function interactWithNearestDoor(token, offsetx = 0, offsety = 0) {
+    // Max distance definition
     let gridSize = canvas.dimensions.size;
     let maxDistance = Infinity;
     let globalMaxDistance = game.settings.get("arms-reach", "globalInteractionDistance");
@@ -102,13 +158,18 @@ document.addEventListener('keyup', evt => {
     var closestDoor = null;
 
     // Find closest door
+    let charCenter = getTokenCenter(token);
+    charCenter.x += offsetx * gridSize;
+    charCenter.y += offsety * gridSize;
+    
     for( let i = 0; i < canvas.controls.doors.children.length ; i++ ) {
       let door = canvas.controls.doors.children[i];
       
-      let dist = getManhattanBetween(door, character);
-      let distInGridUnitys = (dist / gridSize)
+      let dist = getManhattanBetween(door, charCenter);
+      let distInGridUnits = (dist / gridSize) - 0.1;
+
       
-      if ( distInGridUnitys < maxDistance && dist < shortestDistance ) {
+      if ( distInGridUnits < maxDistance && dist < shortestDistance ) {
         closestDoor = door;
         shortestDistance = dist;
       }
@@ -121,19 +182,32 @@ document.addEventListener('keyup', evt => {
       closestDoor._onMouseDown(fakeEvent);
     } else {
       
-      var tokenName = null;
-      if( character.name ) {
-        tokenName = character.name;
-      } else if (character.actor && character.actor.data && character.actor.data.name) {
-        tokenName = character.actor.data.name;
-      }
-      
+      var tokenName = getCharacterName(token);
+
       if (tokenName) iteractionFailNotification("No door was found within " + tokenName + "'s reach" );
       else iteractionFailNotification("No door was found within reach" );
       return;
     }
-	}
-});
+}
+
+// Get token center
+function getTokenCenter(token) {
+    let tokenCenter = {x: token.x , y: token.y };
+    tokenCenter.x += -20 + ( token.w * 0.50 );
+    tokenCenter.y += -20 + ( token.h * 0.50 );
+    return tokenCenter;
+}
+
+// Get chracter name from token
+function getCharacterName(token) {
+  var tokenName = null;
+  if( token.name ) {
+    tokenName = token.name;
+  } else if (token.actor && token.actor.data && token.actor.data.name) {
+    tokenName = token.actor.data.name;
+  }
+  return tokenName;
+}
 
 // Interation fail messages
 function iteractionFailNotification(message) {
@@ -160,5 +234,22 @@ function getSelectedOrOwnedTokens()
 
 // Simple Manhattan Distance between two objects that have x and y attrs.
 function getManhattanBetween(obj1, obj2)  {
+  // console.log("[" + obj1.x + " , " + obj1.y + "],[" + obj2.x + " , " + obj2.y + "]"); // DEBUG
   return Math.abs(obj1.x - obj2.x) + Math.abs(obj1.y - obj2.y);
+}
+
+// Check if active document is the canvas
+function isFocusOnCanvas() {
+  if(   !document.activeElement || 
+        !document.activeElement.attributes ||
+        !document.activeElement.attributes['class'] ||
+        document.activeElement.attributes['class'].value.substr(0,8) !== "vtt game" 
+    ) 
+  { 
+    return false;
+  }
+  else 
+  { 
+    return true;
+  }
 }